requirements.txt
opencv-python
Pillow
PySimpleGUI